Add files via upload
Welcome to my To-Do List project repository!
As part of my internship, I have completed Task 3, which involved creating a to-do list using HTML, CSS, and GitHub. This to-do list allows users to add, edit, and delete tasks, helping them stay organized and productive.

The user interface is designed with HTML for structure and CSS for styling, creating a clean and intuitive layout. The project leverages GitHub for version control and collaboration, demonstrating my ability to use Git for managing code and projects.